Base: Fix file existence check fail for auto load standalone class
When define prefix for standalone class using relate path under include_path
as value, and enable ClassLoader::useIncludePath, ClassLoader will check file
existence (relate to current path) and return.

So, when file to include not exists and useIncludePath is on, need to check if
file exists in include_path too.